+++ lib/src/widgets/cube_animation_widget.dart
@@ -14,6 +14,7 @@ abstract class CubeAnimationWidget<C extends Cube> extends StatelessWidget {
   void dispose() {}
 
   dynamic get initData => null;
+  bool get enableCubeDispose => true;
 
   AnimationController confAnimationController(
     dynamic id, {
@@ -43,6 +44,7 @@ abstract class CubeAnimationWidget<C extends Cube> extends StatelessWidget {
     return CubeBuilderAnimation<C>(
       builder: buildView,
       initData: initData,
+      enableCubeDispose: enableCubeDispose,
       onAction: (cube, data) => onAction(context, cube, data),
       initState: (cube, ticker) {
         tickerContainer.ticker = ticker;